Introduced int64_t HighsInfo type for mip_node_count

+++ check/TestMipSolver.cpp
@@ -167,7 +167,7 @@ TEST_CASE("MIP-integrality", "[highs_test_mip_solver]") {
   if (dev_run) highs.writeSolution("", true);
   REQUIRE(info.objective_function_value == optimal_objective);
 
-  REQUIRE((int)info.mip_node_count == 1);
+  REQUIRE(info.mip_node_count == 1);
   REQUIRE(info.mip_dual_bound == -6);
   REQUIRE(info.mip_gap == 0);
 }
